fix evaluation button
