+++ components/proposals/ProposalPage/components/ProposalEvaluations/components/Review/components/PassFailEvaluation.tsx
@@ -1,79 +1,20 @@
 import { ThumbUpOutlined as ApprovedIcon, ThumbDownOutlined as RejectedIcon } from '@mui/icons-material';
-import PanToolIcon from '@mui/icons-material/PanTool';
 import { Box, Card, Chip, FormLabel, MenuItem, Select, Stack, Typography } from '@mui/material';
 import { usePopupState } from 'material-ui-popup-state/hooks';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 
-import {
-  useAppealProposalEvaluation,
-  useResetEvaluationAppealReview,
-  useResetEvaluationReview,
-  useSubmitEvaluationAppealReview,
-  useSubmitEvaluationReview
-} from 'charmClient/hooks/proposals';
 import { Button } from 'components/common/Button';
 import { UserAndRoleSelect } from 'components/common/DatabaseEditor/components/properties/UserAndRoleSelect';
 import Modal from 'components/common/Modal';
-import MultiTabs from 'components/common/MultiTabs';
 import UserDisplay from 'components/common/UserDisplay';
 import { allMembersSystemRole } from 'components/settings/proposals/components/EvaluationPermissions';
 import { useConfirmationModal } from 'hooks/useConfirmationModal';
-import { useMembers } from 'hooks/useMembers';
-import { useSnackbar } from 'hooks/useSnackbar';
 import { useUser } from 'hooks/useUser';
 import { getActionButtonLabels } from 'lib/proposals/getActionButtonLabels';
 import type { PopulatedEvaluation } from 'lib/proposals/interfaces';
 import { getRelativeTimeInThePast } from 'lib/utils/dates';
 
-export type Props = {
-  hideReviewer?: boolean;
-  proposalId?: string;
-  authors: string[];
-  evaluation: Pick<
-    PopulatedEvaluation,
-    | 'id'
-    | 'completedAt'
-    | 'reviewers'
-    | 'result'
-    | 'isReviewer'
-    | 'actionLabels'
-    | 'requiredReviews'
-    | 'reviews'
-    | 'declineReasonOptions'
-    | 'appealable'
-    | 'appealRequiredReviews'
-    | 'appealReviewers'
-    | 'type'
-    | 'appealedAt'
-    | 'appealedBy'
-    | 'declinedAt'
-    | 'isAppealReviewer'
-  >;
-  refreshProposal?: VoidFunction;
-  confirmationMessage?: string;
-  isCurrent: boolean;
-  archived?: boolean;
-};
-
-function PassFailEvaluationReview({
-  isCurrent,
-  isReviewer,
-  archived,
-  onSubmitEvaluationReview,
-  hideReviewer,
-  reviewerOptions,
-  isSubmittingReview,
-  evaluationReviews = [],
-  requiredReviews,
-  evaluationResult,
-  isResettingEvaluationReview,
-  onResetEvaluationReview,
-  declineReasonOptions,
-  completedAt,
-  actionLabels: _actionLabels,
-  confirmationMessage,
-  isAppealProcess
-}: {
+export type PassFailEvaluationProps = {
   isAppealProcess?: boolean;
   confirmationMessage?: string;
   hideReviewer?: boolean;
@@ -97,7 +38,27 @@ function PassFailEvaluationReview({
   declineReasonOptions: string[];
   completedAt?: Date | null;
   actionLabels?: PopulatedEvaluation['actionLabels'];
-}) {
+};
+
+export function PassFailEvaluation({
+  isCurrent,
+  isReviewer,
+  archived,
+  onSubmitEvaluationReview,
+  hideReviewer,
+  reviewerOptions,
+  isSubmittingReview,
+  evaluationReviews = [],
+  requiredReviews,
+  evaluationResult,
+  isResettingEvaluationReview,
+  onResetEvaluationReview,
+  declineReasonOptions,
+  completedAt,
+  actionLabels: _actionLabels,
+  confirmationMessage,
+  isAppealProcess
+}: PassFailEvaluationProps) {
   const { user } = useUser();
   const currentUserEvaluationReview = evaluationReviews?.find((review) => review.reviewerId === user?.id);
   const [declineReason, setDeclineReason] = useState<string | null>(null);
@@ -124,7 +85,7 @@ function PassFailEvaluationReview({
     if (confirmationMessage) {
       const { confirmed } = await showConfirmation({
         message: confirmationMessage,
-        confirmButton: actionLabels.reject
+        confirmButton: result === 'pass' ? actionLabels.approve : actionLabels.reject
       });
       if (!confirmed) {
         return;
@@ -308,193 +269,6 @@ function PassFailEvaluationReview({
           </Button>
         </Box>
       </Modal>
-    </>
-  );
-}
-
-export function PassFailEvaluation({
-  proposalId,
-  hideReviewer,
-  evaluation,
-  isCurrent,
-  refreshProposal,
-  confirmationMessage,
-  archived,
-  authors
-}: Props) {
-  const [evaluationTab, setEvaluationTab] = useState<number>(0);
-  const { user } = useUser();
-  const { membersRecord } = useMembers();
-  const canAppeal =
-    evaluation.appealable &&
-    !evaluation.appealedAt &&
-    isCurrent &&
-    evaluation.result === 'fail' &&
-    user &&
-    authors.includes(user.id);
-  const { trigger: submitEvaluationReview, isMutating: isSubmittingEvaluationReview } = useSubmitEvaluationReview({
-    proposalId
-  });
-  const { trigger: submitEvaluationAppealReview, isMutating: isSubmittingEvaluationAppealReview } =
-    useSubmitEvaluationAppealReview({ evaluationId: evaluation.id });
-  const { trigger: resetEvaluationReview, isMutating: isResettingEvaluationReview } = useResetEvaluationReview({
-    proposalId
-  });
-  const { trigger: resetEvaluationAppealReview, isMutating: isResettingEvaluationAppealReview } =
-    useResetEvaluationAppealReview({
-      evaluationId: evaluation.id
-    });
-  const { trigger: appealProposalEvaluation, isMutating: isAppealingProposalEvaluation } = useAppealProposalEvaluation({
-    evaluationId: evaluation.id
-  });
-
-  useEffect(() => {
-    if (evaluation.appealedAt) {
-      setEvaluationTab(1);
-    }
-  }, [evaluation.appealedAt]);
-
-  const { showMessage } = useSnackbar();
-  const actionLabels = getActionButtonLabels(evaluation);
-
-  async function onSubmitEvaluationReview(
-    declineReason: string | null,
-    result: NonNullable<PopulatedEvaluation['result']>
-  ) {
-    try {
-      await submitEvaluationReview({
-        evaluationId: evaluation.id,
-        result,
-        declineReasons: declineReason ? [declineReason] : []
-      });
-      refreshProposal?.();
-    } catch (error) {
-      showMessage((error as Error).message, 'error');
-    }
-  }
-
-  async function onSubmitEvaluationAppealReview(
-    declineReason: string | null,
-    result: NonNullable<PopulatedEvaluation['result']>
-  ) {
-    try {
-      await submitEvaluationAppealReview({
-        result,
-        declineReasons: declineReason ? [declineReason] : []
-      });
-      refreshProposal?.();
-    } catch (error) {
-      showMessage((error as Error).message, 'error');
-    }
-  }
-
-  async function onResetEvaluationAppealReview() {
-    resetEvaluationAppealReview().then(refreshProposal);
-  }
-
-  function onResetEvaluationReview() {
-    resetEvaluationReview({
-      evaluationId: evaluation.id
-    }).then(refreshProposal);
-  }
-
-  function onAppealProposalEvaluation() {
-    appealProposalEvaluation().then(refreshProposal);
-  }
-
-  const mainEvaluationReviewComponent = (
-    <PassFailEvaluationReview
-      confirmationMessage={confirmationMessage}
-      hideReviewer={hideReviewer}
-      isCurrent={isCurrent}
-      isReviewer={evaluation.isReviewer}
-      archived={archived}
-      onSubmitEvaluationReview={onSubmitEvaluationReview}
-      onResetEvaluationReview={evaluation.appealedAt ? undefined : onResetEvaluationReview}
-      isResettingEvaluationReview={isResettingEvaluationReview}
-      reviewerOptions={evaluation.reviewers.map((reviewer) => ({
-        group: reviewer.roleId ? 'role' : reviewer.userId ? 'user' : 'system_role',
-        id: (reviewer.roleId ?? reviewer.userId ?? reviewer.systemRole) as string
-      }))}
-      isSubmittingReview={isSubmittingEvaluationReview}
-      evaluationReviews={evaluation.reviews?.filter((review) => !review.appeal)}
-      requiredReviews={evaluation.requiredReviews}
-      evaluationResult={evaluation.appealedAt ? 'fail' : evaluation.result}
-      declineReasonOptions={evaluation.declineReasonOptions}
-      completedAt={evaluation.appealedAt ? evaluation.declinedAt : evaluation.completedAt}
-      actionLabels={actionLabels}
-    />
-  );
-
-  const appealReviewComponent = (
-    <PassFailEvaluationReview
-      isCurrent={isCurrent}
-      isReviewer={evaluation.isAppealReviewer}
-      archived={archived}
-      onSubmitEvaluationReview={onSubmitEvaluationAppealReview}
-      onResetEvaluationReview={evaluation.result === null ? onResetEvaluationAppealReview : undefined}
-      isResettingEvaluationReview={isResettingEvaluationAppealReview}
-      reviewerOptions={
-        evaluation.appealReviewers?.map((reviewer) => ({
-          group: reviewer.roleId ? 'role' : 'user',
-          id: (reviewer.roleId ?? reviewer.userId) as string
-        })) ?? []
-      }
-      isAppealProcess
-      isSubmittingReview={isSubmittingEvaluationAppealReview}
-      evaluationReviews={evaluation.reviews?.filter((review) => Boolean(review.appeal))}
-      requiredReviews={evaluation.appealRequiredReviews ?? 1}
-      declineReasonOptions={evaluation.declineReasonOptions}
-      actionLabels={actionLabels}
-      evaluationResult={evaluation.result}
-      completedAt={evaluation.result !== null ? evaluation.completedAt : undefined}
-    />
-  );
-
-  if (evaluation.appealedAt) {
-    return (
-      <MultiTabs activeTab={evaluationTab} setActiveTab={setEvaluationTab} tabs={['Evaluation', 'Appeal']}>
-        {({ value }) => (
-          <Stack mt={2}>
-            {value === 'Evaluation' && mainEvaluationReviewComponent}
-            {value === 'Appeal' && (
-              <Stack>
-                <Card sx={{ mb: 2 }} variant='outlined'>
-                  <Stack flexDirection='row' gap={1} alignItems='center' justifyContent='center' p={2}>
-                    <PanToolIcon color='warning' />
-                    <Typography variant='body2'>
-                      Appealed by {evaluation.appealedBy ? membersRecord[evaluation.appealedBy]?.username : 'member'}{' '}
-                      {getRelativeTimeInThePast(new Date(evaluation.appealedAt!.toString()))}
-                    </Typography>
-                  </Stack>
-                </Card>
-                {appealReviewComponent}
-              </Stack>
-            )}
-          </Stack>
-        )}
-      </MultiTabs>
-    );
-  }
-
-  return (
-    <>
-      {mainEvaluationReviewComponent}
-      {canAppeal ? (
-        <Stack width='100%' direction='row' justifyContent='flex-end'>
-          <Button
-            sx={{
-              width: 'fit-content',
-              my: 2
-            }}
-            data-test='evaluation-appeal-button'
-            onClick={onAppealProposalEvaluation}
-            loading={isAppealingProposalEvaluation}
-          >
-            Appeal
-          </Button>
-        </Stack>
-      ) : null}
     </>
   );
 }
