+++ package.json
@@ -407,7 +407,8 @@
     "react-dom": "^18.2.0",
     "typescript": "^5.4.2",
     "esbuild": "^0.19.5",
-    "viem": "^2.8.13"
+    "viem": "^2.8.13",
+    "next": "^14.1.1"
   },
   "msw": {
     "workerDirectory": "public"
